chore: refactor character input conversion to InputHelper
