+++ src/Uno.UI.Runtime.Skia.Gtk/Input/GtkKeyboardInputSource.cs
@@ -9,6 +9,7 @@ using Windows.System;
 using Windows.UI.Core;
 using Uno.Foundation.Logging;
 using Windows.Foundation;
+using Uno.UI.Helpers;
 
 namespace Uno.UI.Runtime.Skia.Gtk;
 
@@ -96,44 +97,13 @@ partial class GtkKeyboardInputSource : IUnoKeyboardInputSource
 	{
 		if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
 		{
-			var result = WindowsKeyCodeToUnicode(keyCode);
+			var result = InputHelper.WindowsKeyCodeToUnicode(keyCode);
 			return result.Length > 0 ? result[0] : null; // TODO: supplementary code points
 		}
 
 		var gdkChar = (char)Keyval.ToUnicode(keyVal);
 
 		return gdkChar == 0 ? null : gdkChar;
-	}
-
-	[DllImport("user32.dll")]
-	private static extern bool GetKeyboardState(byte[] lpKeyState);
-
-	[DllImport("user32.dll")]
-	private static extern uint MapVirtualKey(uint uCode, uint uMapType);
-
-	[DllImport("user32.dll")]
-	private static extern IntPtr GetKeyboardLayout(uint idThread);
-
-	[DllImport("user32.dll")]
-	private static extern int ToUnicodeEx(uint wVirtKey, uint wScanCode, byte[] lpKeyState, [Out, MarshalAs(UnmanagedType.LPWStr)] StringBuilder pwszBuff, int cchBuff, uint wFlags, IntPtr dwhkl);
-
-	private static string WindowsKeyCodeToUnicode(uint keyCode)
-	{
-		var keyboardState = new byte[255];
-		var keyboardStateStatus = GetKeyboardState(keyboardState);
-
-		if (!keyboardStateStatus)
-		{
-			return "";
-		}
-
-		var scanCode = MapVirtualKey(keyCode, 0);
-		var inputLocaleIdentifier = GetKeyboardLayout((uint)Environment.CurrentManagedThreadId);
-
-		var result = new StringBuilder();
-		ToUnicodeEx(keyCode, scanCode, keyboardState, result, (int)5, (uint)0, inputLocaleIdentifier);
-
-		return result.ToString();
 	}
 
 	private static VirtualKey ConvertKey(Gdk.Key key)
